Reinstated use of the stop property on maintenance.
The problem wasn't stop, it was that I was 1. using pause for going offline (now using stop), and 2. was never setting stop back to NO. To be clear, pause means that the dispatch_source is paused (the timer is paused while we're performing maintenance) and stop is used to prematurely end the maintenance task, which is important when we go offline and when the program ends. Assuming we even get time to finish what we're doing. Might need to add a high-level -cleanup call to the cache and call that from -applicationWillTerminate: