Delete some entries from the cache if we're over-size.
Currently, in this rough implementation, the size tracked in the maintenance info is much less that the real size on-disk, because the estimates provided by leveldb are not always up to date, especially immediately after inserting a key. So I'll probably now re-do this to just use the size on-disk as the cache size.