Stop trying to track the size of the database in memory.
Rely on the measurements provided by leveldb, even if they are not accurate. It's working well enough to enable keeping the cache within a size bound of around 2x the specified size, before compaction. What we could do is take the size provided by the caller and use a target size around 2/3 of that when thinning the cache, but that value is a guess.