All maintenance data and code now integrated into helper class.
Resetting the cache with -removeAllCachedResponses exposed a timing bug. I was expecting leveldb to handle it until I realized that destroying the database was a class-level operation. For that reason, I first introduced a +destroyDatabase: class method into NULDBDB.

Then it was necessary to synchronize with the maintenance queue to ensure it was stopped, or it would crash when the database was removed from under it. I also found some bugs in the way some variables were being shared with blocks. Now the maintenance class is larger, but it should still avoid leaking due to block references because we stop the maintenance queue with the maintenance block, thereby (I hope) releasing it before we release the maintenance itself.

There are also some places where I was inserting or deleting from the database without being on the correct queue. I cannot perform reads asynchronously on the background queue, however, so those are left unchanged.